All ComparableEntryMaps should store their stuff in doubles rather than in ints, to enable multiplication with doubles

+++ src/dna/profiler/datatypes/ComparableEntryMap.java
@@ -4,7 +4,7 @@ public abstract class ComparableEntryMap implements
 		Comparable<ComparableEntryMap> {
 	public abstract void add(ComparableEntryMap resSecond);
 
-	public abstract void multiplyBy(int factor);
+	public abstract void multiplyBy(double factor);
 
 	public abstract int hashCode();
